feat(video): quality-control v2 — shared bandwidth estimator, playback-rate signal, threshold ramps
Replaces the per-stream AIMD QC with a connection-epoch-anchored
`BandwidthEstimator` shared per direction, a priority `VideoQualityAllocator`,
and capacity-aware caps (`BandwidthCap`, `EncodingCap`, `LayerCap`).

Inbound signal switches from capture-time drift to **playback rate**
(source-clock ms drained per wall-clock ms at the present stage), sampled
at 1 Hz with EMA α = 0.3 in `present-mstg` / `present-canvas`. Drop penalty
filters to receiver-side `FrameDropStage` 61–90 only — sender + server
drops are intentional pacing and don't reflect downstream health.

Penalties are now linear ramps from `Ok` (0) to `Bad` (1):
- `PlaybackRateOk = 0.90`, `PlaybackRateBad = 0.00`
- `DropOkReceiver = 0.20`, `DropBadReceiver = 0.50`
- `DropOkSender = 0.20`, `DropBadSender = 0.50`

Drops `QualityReductionRequested` (never set true; entire QR branch removed
from wire type, classifier, allocator, and modal).

Diagnostics modal refactored into 3 tabs (Inbound / Outbound / Settings)
with text-selectable body and a Copy button using
`navigator.clipboard.writeText`.

Docs updated to current state: `docs/live-video/08-quality-control.md`
and `docs/plans/video-quality-control-v2.md`.

+++ src/dotnet/Api.Contracts/Streaming/Quality/RecordingQuality.cs
@@ -48,10 +48,10 @@ public sealed record RecorderStats(
     bool IsPeerConnected,
     // Cumulative per-stage drop counts since the recording run started.
     // Powers the per-stream FPS breakdown in video diagnostics.
-    IReadOnlyDictionary<FrameDropStage, long> DropTrace,
+    IReadOnlyDictionary<FrameDropStage, int> DropTrace,
     // Cumulative bundles successfully shipped to the wire since the run
     // started. Powers FPS (= delta-per-second) for the diagnostics row.
-    long BundlesShipped,
+    int BundlesShipped,
     // Cumulative encoded bytes (sum across layers). Drives the outbound
     // "kbps" display in the diagnostics modal.
     long BytesEncoded)
@@ -59,6 +59,6 @@ public sealed record RecorderStats(
     public static RecorderStats Empty { get; } =
         new(0, 0, 0, IsConnected: false, IsPeerConnected: false, EmptyDropTrace, 0, 0);
 
-    private static readonly IReadOnlyDictionary<FrameDropStage, long> EmptyDropTrace
-        = new Dictionary<FrameDropStage, long>();
+    private static readonly IReadOnlyDictionary<FrameDropStage, int> EmptyDropTrace
+        = new Dictionary<FrameDropStage, int>();
 }
